renamed folder and added travis
